grit: tool to copy commits between the monorepo and Github
Summary:
Grit is a very simple tool to replace fbshipit, which is overly complicated
for our purposes. Grit packages just the functionality used by GRAIL into
a single Go binary.

Grit tracks synchronization points like fbshipit does, and thus is compatible
with it (backwards and forwards).

The current implementation does not attempt to recover from errors, but
it is simple enough in operation and easy to understand, and so it's quite
easy to apply manual fixes where necessary. It also provides us with a basis
for adding features as needed.

Reviewers: ysaito, sbagaria

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D21607

fbshipit-source-id: 5770274