grit: skip unapplicable commits when synchronizing two repositories
Summary:
Some files like Bazel BUILD files as well as go.{mod,sum} files are
routinely updated as part of seemingly unrelated commits. For
example, at GRAIL we may modify bigslice/go.mod when a change from
base is imported from GitHub. The commit message contains the
"shipit" tag that grit looks for, and thus will synchronize to the
wrong commit.

The solution is to apply the rewrite rules while looking for the
synchronizing commit, so that we can skip commits that would have no
effect on the particular destination.

Reviewers: ysaito, sbagaria

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D38148

fbshipit-source-id: 1f7adca